Binded archive format to repository backend To satisfy more complex import/export scenarios, like importing an archive that contains file keys generated differently from the default sha256, the archive_format has been binded to the repository backend class, this way the import/export format is relative to the mounted storage backend.
Before this change importing an archive with a different key format
wasn't possible since the verdi command was forcing the archive format
to the only one availabe 'ArchiveFormatSqlZip' that was forcing imported
archives to have the sha256 key format even if it was different from the
key format of the mounted storage backend.

For example: using aiida-s3 storage backend it wasn't possible to import
an exported archive with uuid4 key_format.

An additional property `maintain_keys` has been added to the ArchiveFormatAbstract that manages if the imported file keys have to be regenerated or to use the same from the archive.